Sling export of a hierarchy of pages -- HierarchyPage and Page merged -- initial implementation (fields type selection) for policy fields -- unit test merge and adaptation
