+++ bundles/core/pom.xml
@@ -150,6 +150,7 @@
                                         <!-- private enum; cannot test valueOf / values -->
                                         <exclude>**/SocialMediaHelperImpl$WebsiteMetadata$Type.class</exclude>
                                         <exclude>**/AdaptiveImageServlet$Source.class</exclude>
+                                        <exclude>**/HierarchyComponentContextWrapper.class</exclude>
                                     </excludes>
                                 </configuration>
                             </execution>
