+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/internal/models/v2/PageImplTest.java
@@ -20,14 +20,20 @@ import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
+import org.apache.sling.api.SlingHttpServletRequest;
+import org.apache.sling.api.resource.Resource;
+import org.apache.sling.models.factory.ModelFactory;
+import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.osgi.framework.Version;
 import org.powermock.reflect.Whitebox;
 
+import com.adobe.cq.export.json.hierarchy.HierarchyNodeExporter;
 import com.adobe.cq.wcm.core.components.Utils;
 import com.adobe.cq.wcm.core.components.models.NavigationItem;
 import com.adobe.cq.wcm.core.components.models.Page;
@@ -41,12 +47,16 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.when;
 
 public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.models.v1.PageImplTest {
 
     private static final String TEST_BASE = "/page/v2";
     private static final String REDIRECT_PAGE = ROOT + "/redirect-page";
+
+    private static final String PAGE_CHILD = PAGE + "/child";
 
     private static ClientLibrary mockClientLibrary;
 
@@ -85,6 +95,39 @@ public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.mode
         assertArrayEquals(new String[] {"coretest.product-page"}, page.getClientLibCategoriesJsBody());
         assertEquals("product-page", page.getTemplateName());
         Utils.testJSONExport(page, Utils.getTestExporterJSONPath(TEST_BASE, PAGE));
+    }
+
+    @Test
+    public void testChildren() {
+        Page page = getPageUnderTest(PAGE);
+        ModelFactory modelFactory = Mockito.mock(ModelFactory.class);
+        Whitebox.setInternalState(page, "modelFactory", modelFactory);
+
+        Page childPage = getPageUnderTest(PAGE_CHILD);
+        when(modelFactory.getModelFromWrappedRequest(any(SlingHttpServletRequest.class), any(Resource.class), eq(Page.class))).thenReturn(childPage);
+
+        Map<String, ? extends HierarchyNodeExporter> children = page.getExportedChildren();
+        Assert.assertEquals(2, children.size());
+        Assert.assertEquals(childPage, children.get(PAGE_CHILD));
+
+    }
+    @Test
+    public void testGetRootUrl() {
+
+        // no parent
+        Page rootPage = getPageUnderTest(PAGE);
+        Assert.assertEquals(CONTEXT_PATH + PAGE + ".model.json", rootPage.getRootUrl());
+
+        // with parent
+        Page childPage = getPageUnderTest(PAGE_CHILD);
+        Assert.assertEquals(CONTEXT_PATH + PAGE + ".model.json", childPage.getRootUrl());
+    }
+
+    @Test
+    public void testGetRootModel() {
+        // no parent
+        Page page = getPageUnderTest(PAGE);
+        Assert.assertEquals(page, page.getRootModel());
     }
 
     @Test(expected = UnsupportedOperationException.class)
